backend/requirements.txt
flask
python-dotenv
flask-restful 